chat_sync/wecom/transformer/text_content_field_transformer.go
package transformer

import (
	"errors"
	"github.com/yrosukedev/chat_record_sync/chat_sync/business"
	"github.com/yrosukedev/chat_record_sync/chat_sync/wecom"
)

type TextContentFieldTransformer struct {
}

func NewTextContentFieldTransformer() *TextContentFieldTransformer {
	return &TextContentFieldTransformer{}
}

func (t *TextContentFieldTransformer) Transform(wecomRecord *wecom.ChatRecord, chatRecord *business.ChatRecord) (updatedChatRecord *business.ChatRecord, err error) {
	if wecomRecord == nil {
		return nil, errors.New("wecomRecord can't be nil")
	}

	updatedChatRecord = t.copyInputIfNeeded(chatRecord)

	if wecomRecord.Text != nil {
		updatedChatRecord.Content = wecomRecord.Text.Content
	}

	return updatedChatRecord, nil
}

func (t *TextContentFieldTransformer) copyInputIfNeeded(chatRecord *business.ChatRecord) *business.ChatRecord {
	updatedChatRecord := &business.ChatRecord{}
	if chatRecord != nil {
		*updatedChatRecord = *chatRecord
	}
	return updatedChatRecord
}
